+++ components/App.js
@@ -624,10 +624,6 @@ class App extends Component {
             if (button === 0) {
                 if (board.get(vertex) === 0) {
                     this.makeMove(vertex)
-
-                    if (this.attachedEngineControllers.some(x => x != null)) {
-                        setTimeout(() => this.startGeneratingMoves(), setting.get('gtp.move_delay'))
-                    }
                 } else if (vertex in board.markups
                 && board.markups[vertex][0] === 'point'
                 && setting.get('edit.click_currentvertex_to_remove')) {
@@ -723,7 +719,7 @@ class App extends Component {
             if (!board.hasVertex(nextVertex)) return this.setMode('play')
 
             if (helper.vertexEquals(vertex, nextVertex)) {
-                this.makeMove(vertex, {player: 'B' in nextNode ? 1 : -1})
+                this.makeMove(vertex, {player: 'B' in nextNode ? 1 : -1, sendToEngine: false})
             } else {
                 if (board.get(vertex) !== 0
                 || this.state.blockedGuesses.some(v => helper.vertexEquals(v, vertex)))
@@ -750,7 +746,7 @@ class App extends Component {
         this.events.emit('vertexClick')
     }
 
-    makeMove(vertex, {player = null, clearUndoPoint = true} = {}) {
+    makeMove(vertex, {player = null, clearUndoPoint = true, sendToEngine = true} = {}) {
         if (!['play', 'autoplay', 'guess'].includes(this.state.mode)) {
             this.closeDrawer()
             this.setMode('play')
@@ -914,6 +910,12 @@ class App extends Component {
         // Emit event
 
         this.events.emit('moveMake', {pass, capture, suicide, ko, enterScoring})
+
+        // Send command to engine
+
+        if (sendToEngine && this.attachedEngineControllers.some(x => x != null)) {
+            setTimeout(() => this.startGeneratingMoves(), setting.get('gtp.move_delay'))
+        }
     }
 
     makeResign({player = null, setUndoPoint = true} = {}) {
@@ -925,7 +927,7 @@ class App extends Component {
         if (setUndoPoint) this.setUndoPoint('Undo Resignation')
         rootNode.RE = [`${color}+Resign`]
 
-        this.makeMove([-1, -1], {player, clearUndoPoint: false})
+        this.makeMove([-1, -1], {player, clearUndoPoint: false, sendToEngine: false})
         this.makeMainVariation(...this.state.treePosition, {setUndoPoint: false})
 
         this.events.emit('resign', {player})
@@ -2106,7 +2108,7 @@ class App extends Component {
                 return
             }
 
-            this.makeMove(vertex, {player: sign})
+            this.makeMove(vertex, {player: sign, sendToEngine: false})
             this.engineBoards[playerIndex] = gametree.getBoard(...this.state.treePosition)
 
             if (otherController != null && !helper.vertexEquals(vertex, [-1, -1])) {